requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
dj-database-url==0.4.0
Django==1.9.2
django-filter==0.12.0
djangorestframework==3.3.2
gunicorn==19.4.5
kombu==3.0.33
Markdown==2.6.5
Pillow==3.1.1
psycopg2==2.6.1
pytz==2015.7
redis==2.10.5
reportlab==3.3.0
requests==2.9.1
whitenoise==2.0.6
